Transfer fetchLoggedInStatus function into FetchMethods.js file, update import statement in various components
